requirements.txt
click
python-telegram-bot==13.15
retry
tzlocal==2.1
bs4
tweepy_authlib
